+++ tests/testthat/test_convertAnnotation.R
@@ -1,4 +1,5 @@
-# Test main convertAnnotation function ====
+## Test main convertAnnotation function
+
 data(modifications)